+++ crates/defguard_core/src/enterprise/grpc/desktop_client_mfa.rs
@@ -22,6 +22,7 @@ use crate::{
     },
     events::{BidiRequestContext, BidiStreamEvent, BidiStreamEventType, DesktopClientMfaEvent},
     grpc::{proxy::client_mfa::ClientMfaServer, utils::parse_client_ip_agent},
+    mfa_engine::authorize::emit_event,
 };
 
 impl ClientMfaServer {
@@ -127,17 +128,20 @@ impl ClientMfaServer {
             Ok(url) => url,
             Err(status) => {
                 self.delete_mfa_session(session).await?;
-                self.emit_event(BidiStreamEvent {
-                    context,
-                    event: BidiStreamEventType::DesktopClientMfa(Box::new(
-                        DesktopClientMfaEvent::Failed {
-                            location: location.clone(),
-                            device: device.clone(),
-                            method,
-                            message: "provided invalid callback URL".to_owned(),
-                        },
-                    )),
-                })?;
+                emit_event(
+                    &self.bidi_event_tx,
+                    BidiStreamEvent {
+                        context,
+                        event: BidiStreamEventType::DesktopClientMfa(Box::new(
+                            DesktopClientMfaEvent::Failed {
+                                location: location.clone(),
+                                device: device.clone(),
+                                method,
+                                message: "provided invalid callback URL".to_owned(),
+                            },
+                        )),
+                    },
+                )?;
                 return Err(status);
             }
         };
@@ -160,17 +164,22 @@ impl ClientMfaServer {
                 if claims_user.id != user.id {
                     info!("User {claims_user} tried to use OIDC MFA for another user: {user}");
                     self.delete_mfa_session(session).await?;
-                    self.emit_event(BidiStreamEvent {
-                        context,
-                        event: BidiStreamEventType::DesktopClientMfa(Box::new(
-                            DesktopClientMfaEvent::Failed {
-                                location: location.clone(),
-                                device: device.clone(),
-                                method,
-                                message: format!("user {claims_user} tried to use OIDC MFA for another user: {user}")
-                            },
-                        )),
-                    })?;
+                    emit_event(
+                        &self.bidi_event_tx,
+                        BidiStreamEvent {
+                            context,
+                            event: BidiStreamEventType::DesktopClientMfa(Box::new(
+                                DesktopClientMfaEvent::Failed {
+                                    location: location.clone(),
+                                    device: device.clone(),
+                                    method,
+                                    message: format!(
+                                        "user {claims_user} tried to use OIDC MFA for another user: {user}"
+                                    ),
+                                },
+                            )),
+                        },
+                    )?;
                     return Err(Status::unauthenticated("unauthorized"));
                 }
                 info!(
@@ -181,17 +190,20 @@ impl ClientMfaServer {
             Err(err) => {
                 info!("Failed to verify OIDC code: {err}");
                 self.delete_mfa_session(session).await?;
-                self.emit_event(BidiStreamEvent {
-                    context,
-                    event: BidiStreamEventType::DesktopClientMfa(Box::new(
-                        DesktopClientMfaEvent::Failed {
-                            location: location.clone(),
-                            device: device.clone(),
-                            method,
-                            message: format!("failed to verify OIDC code: {err}"),
-                        },
-                    )),
-                })?;
+                emit_event(
+                    &self.bidi_event_tx,
+                    BidiStreamEvent {
+                        context,
+                        event: BidiStreamEventType::DesktopClientMfa(Box::new(
+                            DesktopClientMfaEvent::Failed {
+                                location: location.clone(),
+                                device: device.clone(),
+                                method,
+                                message: format!("failed to verify OIDC code: {err}"),
+                            },
+                        )),
+                    },
+                )?;
                 return Err(Status::unauthenticated("unauthorized"));
             }
         }
